pyproject.toml
[tool.poetry]
name = "allset"
version = "0.1.0"
description = "active learning via influence functions"
authors = ["Pietro Lesci <pietrolesci@outlook.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "energizer" },
]

[tool.poetry.dependencies]
python = "^3.8"
lightning = "^2.0.0"
transformers = "^4.28.1"
datasets = "^2.12.0"
pandas = "^2.0.1"
sentence-transformers = "^2.2.2"
hydra-core = "^1.3.1"
omegaconf = "^2.3.0"
colorlog = "^6.7.0"
fastapi = "^0.86.0"
torch-optimizer = "^0.3.0"
matplotlib = "^3.6.3"
seaborn = "^0.12.2"
srsly = "^2.4.5"
hydra-colorlog = "^1.2.0"
hnswlib = {git = "https://github.com/nmslib/hnswlib.git"}
tensorboard = "^2.12.0"
torch-tb-profiler = "^0.4.1"
tabulate = "^0.9.0"
openpyxl = "^3.1.1"
hydra-joblib-launcher = "^1.2.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
nbqa = "^1.6.1"
ipykernel = "^6.20.2"
ipywidgets = "7.7.1"
isort = "^5.11.4"
pytest = "^7.2.1"
nvitop = "^1.0.0"


[tool.black]
line-length = 120
skip-string-normalization = false
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 120
src = ["src", "test"]
target-version = "py38"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
